+++ tacacs-plus/src/context.rs
@@ -56,9 +56,9 @@ pub struct ContextBuilder {
 
 impl ContextBuilder {
     /// Creates a new builder with default values for the various fields.
-    pub fn new(user: &str) -> Self {
+    pub fn new(user: String) -> Self {
         Self {
-            user: user.to_owned(),
+            user,
             port: String::from("rust_client"),
             remote_address: String::from("tacacs_plus_rs"),
             privilege_level: Default::default(),
@@ -67,19 +67,19 @@ impl ContextBuilder {
     }
 
     /// Sets the port of the resulting context.
-    pub fn port(mut self, port: String) -> Self {
+    pub fn port(&mut self, port: String) -> &mut Self {
         self.port = port;
         self
     }
 
     /// Sets the remote address of the resulting context.
-    pub fn remote_address(mut self, remote_address: String) -> Self {
+    pub fn remote_address(&mut self, remote_address: String) -> &mut Self {
         self.remote_address = remote_address;
         self
     }
 
     /// Sets the privilege level of the resulting context.
-    pub fn privilege_level(mut self, privilege_level: PrivilegeLevel) -> Self {
+    pub fn privilege_level(&mut self, privilege_level: PrivilegeLevel) -> &mut Self {
         self.privilege_level = privilege_level;
         self
     }
@@ -87,17 +87,17 @@ impl ContextBuilder {
     /// Sets the authentication method of the resulting context.
     ///
     /// Note that this field is ignored in an authentication session.
-    pub fn auth_method(mut self, method: AuthenticationMethod) -> Self {
+    pub fn auth_method(&mut self, method: AuthenticationMethod) -> &mut Self {
         self.authentication_method = Some(method);
         self
     }
 
     /// Consumes this builder and turns it into a [`SessionContext`].
-    pub fn build(self) -> SessionContext {
+    pub fn build(&self) -> SessionContext {
         SessionContext {
-            user: self.user,
-            port: self.port,
-            remote_address: self.remote_address,
+            user: self.user.clone(),
+            port: self.port.clone(),
+            remote_address: self.remote_address.clone(),
             privilege_level: self.privilege_level,
             authentication_method: self.authentication_method,
         }